Add basic WASM integration in editor
Add a endive-wasm crate which will serve as the WebAssembly interface
for Endive. Use wasm-pack to compile it and import it in the editor.

WebAssembly support in Rust requires the use of a special target, so
modify the Nix shell definition to add this target. To do that, we
switch to oxalica's Rust overlay.