Add type for inductive type families to the kernel
The goal is to move from the current model where
lambda terms contain duplicate definitions of
inductive type families to a model where the
families are known upfront and can be refenced in
lambda terms.

The inductive type families now need to be
parameterized to have the same flexibility as
before.

The strict positivity constraint is now directly
encoded in the host type system.